Stop carrying mistakes from C's legacy: change the precedence of operators.
To quote DMR: "Today, it seems that it would have been preferable to move the
relative precedences of & and ==, and thereby simplify a common C idiom: to
test a masked value against another value, one must write:
```if ((a&mask) == b) ...```
where the inner parentheses are required but easily forgotten."

The same paper also explains that this was carrying forward B's legacy where
& and && weren't seperate operators. Since we have && and || to compound
comparisons in conditionals, this is completely unnecessary, thus this commit
brings the precedence of operators in-line with Rust.

In addition let's make all the comparison operators <, <=, >, >=, == and !=
have the same precedence, and make all of them non-associative. Their
associativity was questionable anyways, as a construct like "3 < x < 5" in C,
doesn't actually do what one would expect.

Thus the current operator precdences and associativities are as follows:

Op              | Associativity
----------------+----------------
as (cast)       | left-to-right
----------------+----------------
* / %           | left-to-right
----------------+----------------
+ -             | left-to-right
----------------+----------------
<< >>           | left-to-right
----------------+----------------
&               | left-to-right
----------------+----------------
^               | left-to-right
----------------+----------------
|               | left-to-right
----------------+----------------
< <= > >= == != | non-associative
----------------+----------------
&&              | left-to-right
----------------+----------------
||              | left-to-right

NOTE: previous versions of the parser associated operators right-to-left.
This actually wasn't desired and is now fixed.