Require parenthesis around "if" and "while" conditions.
A premature choice like this is what happens when one blindly follows "good ideas"
from other languages.

Not requiring said parenthesis and introducing Rust-like record literals actually
made the current grammar ambigous (at least with 1 token lookahead), consider:
```
if a {
}
```

This could either be parsed as,
```if (a) {}```
where the expression of the if statement is the identifierr "a", or
```if (a {})```,
where "a {}" is a record literal.

Rust resolves this ambiguity by not allowing struct literals in conditionals like
this unless parenthized. While this makes the language unambigous, writing a grammar
and parser for this requires making the parser context sensitive, or duplicating
the sub-grammar of expressions based on what context they appear in. Neither of
which option I consider acceptable, thus the indroduction of the parenthesis.